requirements.txt
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==5.12.4
twilio==8.11.1
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

